# openapi.yaml
openapi: 3.0.3
info:
  title: C2PA Soft Binding Resolution API
  description: >
    Implementation of the C2PA Soft Binding Resolution API
    (https://spec.c2pa.org/specifications/specifications/2.4/softbinding/Decoupled.html).
    Allows clients to store C2PA Manifest Stores, associate them with soft
    binding values (watermarks or fingerprints), and recover those manifests
    from a previously detected binding.
  version: '2.4'
servers:
  - url: /v1
security:
  - oauth2:
      - fetch:manifests
...
paths:
...
  /manifests:
    post:
      tags: [Store]
      summary: Ingest a C2PA Manifest Store
      operationId: postManifest
      security:
        - oauth2:
            - store:manifests
      parameters:
        - name: returnReceipt
          in: query
          description: If true, generate and return a repository receipt for the manifest.
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/c2pa:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Manifest stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  manifestId:
                    type: string
                    example: 'urn:c2pa:12345678-1234-1234-1234-123456789abc'
                  receipt:
                    $ref: '#/components/schemas/Receipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/ServiceFailure'
  /bindings:
    post:
      tags: [Store]
      summary: Associate a soft binding value with a stored manifest
      operationId: createBinding
      security:
        - oauth2:
            - store:bindings
      requestBody:
        $ref: '#/components/requestBodies/Binding'
      responses:
        '204':
          description: Binding created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '404':
          description: Soft binding id or C2PA Manifest id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
    put:
      tags: [Store]
      summary: Replace the manifest a soft binding value points to
      operationId: updateBinding
      security:
        - oauth2:
            - store:bindings
      requestBody:
        $ref: '#/components/requestBodies/Binding'
      responses:
        '204':
          description: Binding updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '404':
          description: Soft binding value not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
  /manifests/{manifestId}:
    get:
      tags: [Fetch]
      summary: Retrieve a stored C2PA Manifest Store
      operationId: getManifest
      parameters:
        - $ref: '#/components/parameters/manifestId'
      responses:
        '200':
          description: The C2PA Manifest Store
          content:
            application/c2pa:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: C2PA Manifest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
    delete:
      tags: [Store]
      summary: Remove a manifest and its associated bindings
      operationId: deleteManifest
      security:
        - oauth2:
            - store:manifests
      parameters:
        - $ref: '#/components/parameters/manifestId'
      responses:
        '204':
          description: Manifest deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: C2PA Manifest Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServiceFailure'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth2 client credentials. Configure the token URL for your identity provider.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth/token
          scopes:
            fetch:manifests: Query and fetch manifests and receipts
            store:manifests: Store and delete manifests
            store:bindings: Create and update soft bindings
  parameters:
...
    manifestId:
      name: manifestId
      in: path
      required: true
      description: Identifier of a stored C2PA Manifest Store
      schema:
        type: string
  requestBodies:
    Binding:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [bindingValue, manifestId]
            properties:
              bindingValue:
                type: string
                description: Base64-encoded soft binding value
              manifestId:
                type: string
                description: Identifier of a previously stored C2PA Manifest Store
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The authenticated principal lacks the required authorization scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Request body exceeds the configured size limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceFailure:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Receipt:
      type: object
      properties:
        '@context':
          type: object
          properties:
            c2pa:
              type: string
            receipt:
              type: string
        '@type':
          type: string
          example: org.c2pa.manifest-receipt
        repository:
          type: object
          properties:
            uri:
              type: string
            manifestId:
              type: string
        anchor:
          type: object
          properties:
            uri:
              type: string
            proof:
              type: object
              properties:
                alg:
                  type: string
                value:
                  type: string
        verified:
          type: boolean
        error:
          type: string